refactor(core): move WorkflowStore trait + DTOs to assay-core
Trait body and shared DTOs (ApiKeyRecord, NamespaceRecord,
NamespaceStats, QueueStats) now live in assay-core. assay-workflow's
store module shrinks to module declarations + re-exports. Existing
crate::store::WorkflowStore and crate::store::<DTO> import paths
resolve unchanged via re-export.
