+++ crates/assay-core/Cargo.toml
@@ -13,3 +13,6 @@ anyhow = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 utoipa = { version = "5", features = ["axum_extras"] }
+
+# Stream trait for the push-subscription store methods
+futures-core = "0.3"
