+++ crates/assay-core/src/lib.rs
@@ -3,6 +3,8 @@
 //! Consumers: `assay-workflow`, `assay-auth`, `assay-engine`.
 //! Backend impls live in the domain crates behind Cargo features.
 
+pub mod store;
 pub mod types;
 
+pub use store::WorkflowStore;
 pub use types::*;
